Move version tests into another file. Bring back all the tests.
