+++ package.json
@@ -32,7 +32,7 @@
     "mocha-light": "mocha --node-env=test --config .mocharc.json --exclude \"./dist/test/integration/compute.test.js\"",
     "test": "npm run lint && npm run test:unit:cover && npm run test:integration:cover",
     "test:unit": "npm run build-tests && npm run mocha \"./dist/test/unit/**/*.test.js\"",
-    "test:integration": "npm run build-tests && npm run mocha \"./dist/test/integration/database.test.js\"",
+    "test:integration": "npm run build-tests && npm run mocha \"./dist/test/integration/**/*.test.js\"",
     "test:indexer": "npm run build-tests && npm run mocha \"./dist/test/integration/indexer.test.js\"",
     "test:integration:light": "npm run build-tests && npm run mocha-light \"./dist/test/integration/**/*.test.js\"",
     "test:unit:cover": "nyc --report-dir coverage/unit npm run test:unit",
